fix indexing

+++ src/cross_validation_anditi_v3.py
@@ -266,7 +266,7 @@ def plot_worst_ns(cwd, num_images=10):
             path = f"/mnt/sdb/agorup/school_mapping/satellite_images/large/VNM/non_school/{row['UID']}.jpeg"
             image = Image.open(path).convert("RGB")
 
-            plt.subplot(2, num_images / 2, i + 1)
+            plt.subplot(2, int(num_images / 2), i + 1)
             plt.imshow(image)
             plt.axis('off')
         plt.savefig(os.path.join(cwd, f"{num_images}_worst_ns_train_{i}.png"))
@@ -290,7 +290,7 @@ def plot_worst_ns(cwd, num_images=10):
             path = f"/mnt/sdb/agorup/school_mapping/satellite_images/large/VNM/non_school/{row['UID']}.jpeg"
             image = Image.open(path).convert("RGB")
 
-            plt.subplot(2, num_images / 2, i + 1)
+            plt.subplot(2, int(num_images / 2), i + 1)
             plt.imshow(image)
             plt.axis('off')
         plt.savefig(os.path.join(cwd, f"{num_images}_worst_ns_val_{i}.png"))
